https://ru.yougile.com/team/d193f0b54f05/#KAN-12

+++ app/schemas/admin.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, ConfigDict, Field
 
@@ -17,7 +18,7 @@ class AdminLocationRead(LocationBase):
     model_config = ConfigDict(from_attributes=True)
 
     id: int
-    distance_to_city_km: int | None = Field(default=None, ge=0)
+    distance_to_city_km: Decimal | None = Field(default=None, ge=0)
     city: str = Field(min_length=1, max_length=150)
     region: str | None = Field(default=None, max_length=150)
     country: str | None = Field(default=None, max_length=150)
